+++ processor/src/test/fixtures/bad_input/com/example/injector/includes/ExplicitIncludesOfEnclosedInjectableModel.java
@@ -1,0 +1,15 @@
+package com.example.injector.includes;
+
+import sting.Injectable;
+import sting.Injector;
+
+@Injector( includes = ExplicitIncludesOfEnclosedInjectableModel.MyModel.class )
+interface ExplicitIncludesOfEnclosedInjectableModel
+{
+  MyModel getMyModel();
+
+  @Injectable
+  class MyModel
+  {
+  }
+}