+++ processor/src/test/fixtures/bad_input/com/example/fragment/EnclosedAnnotationFragmentModel.java
@@ -1,0 +1,16 @@
+package com.example.fragment;
+
+import sting.Fragment;
+
+@Fragment
+public interface EnclosedAnnotationFragmentModel
+{
+  default Runnable provideRunnable()
+  {
+    return null;
+  }
+
+  @interface Foo
+  {
+  }
+}