+++ processor/src/test/fixtures/bad_input/com/example/fragment/EnclosedEnumFragmentModel.java
@@ -1,0 +1,16 @@
+package com.example.fragment;
+
+import sting.Fragment;
+
+@Fragment
+public interface EnclosedEnumFragmentModel
+{
+  default Runnable provideRunnable()
+  {
+    return null;
+  }
+
+  enum Foo
+  {
+  }
+}