+++ processor/src/test/fixtures/bad_input/com/example/fragment/EnclosedInterfaceFragmentModel.java
@@ -1,0 +1,16 @@
+package com.example.fragment;
+
+import sting.Fragment;
+
+@Fragment
+public interface EnclosedInterfaceFragmentModel
+{
+  default Runnable provideRunnable()
+  {
+    return null;
+  }
+
+  interface Foo
+  {
+  }
+}