+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.10](https://github.com/sting-ioc/sting/tree/v0.10) (2020-03-25) · [Full Changelog](https://github.com/sting-ioc/sting/compare/v0.09...v0.10)
+
+Changes in this release:
 
 * Generate an error if an `@Injector` annotated type encloses a type that is not annotated with either `@Injectable` or `@Fragment`.
 * Eliminate a crash when a primitive type dependency was missing and the processor attempted to auto-discover the type.